Add world.js and create class
