+++ src/index.js
@@ -1,6 +1,6 @@
 import Phaser from 'phaser';
-import { WorldScene } from './modules/game';
-import BootScene from './modules/game'
+import WorldScene from './modules/world';
+import BootScene from './modules/game';
 import BattleScene from './modules/battle';
 import { UIScene } from './modules/interface';
